+++ CHANGELOG.md
@@ -115,3 +115,4 @@
 * Generalize `$` up to a method of `Divariant` typeclass
 * Put concrete data structures to `Specific` submodule
 * Move `Nonempty` type family to separated module
+* Define `Cartesian` type class